requirements.txt
async-timeout==3.0.1
discord.py==1.3.2
dnspython==1.16.0
pymongo==3.10.1
pyEX==0.2.1
requests~=2.23.0
us~=2.0.1
requests-cache==0.5.2